[Scene]: 'spinning-gears' scene is moved to React 'SpinningGearsScene' component.
Changelog:
- old features is restored in new 'scene' component;
- 'spinning-gears/gear.class.js' is updated;
- old 'scene.js' file is deleted now.
